816* Various: Changes to compile on 3.6.8/ARM/RaspberryPi kernel.           ARM is still no good to run. This build *should* compile 	  apart from the very last section on divmod64.c processing, 	  but am releasing to keep current. Theres still a lot 	  of execution time work to do.

+++ driver/dtrace_asm.c
@@ -141,7 +141,7 @@ dtrace_casptr(void *target, void *cmp, void *new)
 uint32_t
 dtrace_cas32(uint32_t *target, uint32_t cmp, uint32_t new)
 {
-	return cmpxchg((void **) target, (void **) cmp, (void **) new);
+	return (uint32_t) cmpxchg((void **) target, (void **) cmp, (void **) new);
 }
 void *
 dtrace_casptr(void *target, void *cmp, void *new)
